v0.3
Added ALL the logic to fetch, pass and display the data about forecasting for the week.
App is still missing the correct data for the "Month" box, but the feature is coming in v0.6 release.
All the data displayed is correct and it flows flawlessly throughout the components' annidation.
Added the two static components "Search" and "Localization".
Slightly reorganized the hierarchy of components: now the "week/month" choice is handled by a wrapper that passes the current display mode to its child, telling him wether to display Week or Month data.

The app throws a warning regarding the "key" prop of the table components, a fix for it is coming soon.