v0.8
Adjusted the layout according to the requests: AddCity component is now centered in the right part of the layout.
Alternative cities now have a background color according to the weather they're reporting (grey for clouds, blue for clear sky, black/dark grey for tempests, grey-light blue gradient for snow and hail and so on.
App is now fully responsive: tweaked the past layout to stretch accordingly to small screens.

TO DO: Find a reliable API to fetch the monthly data, since OpenWeather does not provide a free plan for this kind of forecasting.